+++ src/main/typescript/applications/actor/wvActorSheet.ts
@@ -3,6 +3,7 @@ import {
   CONSTANTS,
   EquipmentSlot,
   HANDLEBARS,
+  isApparelSlot,
   isEquipmentSlot,
   isPhysicalItemType,
   isSkillName,
@@ -140,6 +141,13 @@ export default class WvActorSheet extends ActorSheet {
         );
       });
     sheetForm
+      .querySelectorAll("button[data-action=unequip]")
+      .forEach((element) => {
+        element.addEventListener("click", (event) =>
+          this.onClickUnequipItem(event)
+        );
+      });
+    sheetForm
       .querySelectorAll("button[data-weapon-attack-name]")
       .forEach((element) => {
         element.addEventListener("click", (event) =>
@@ -165,12 +173,12 @@ export default class WvActorSheet extends ActorSheet {
         ? this.toSheetWeapon(actorReadiedItem)
         : actorReadiedItem?.toObject(false) ?? null;
     const armor =
-      this.actor.armor instanceof Apparel
-        ? this.toSheetApparel(this.actor.armor)
+      this.actor.armorApparel instanceof Apparel
+        ? this.toSheetApparel(this.actor.armorApparel)
         : null;
     const clothing =
-      this.actor.clothing instanceof Apparel
-        ? this.toSheetApparel(this.actor.clothing)
+      this.actor.clothingApparel instanceof Apparel
+        ? this.toSheetApparel(this.actor.clothingApparel)
         : null;
     const eyes =
       this.actor.eyesApparel instanceof Apparel
@@ -235,7 +243,13 @@ export default class WvActorSheet extends ActorSheet {
       sheet: {
         bounds: CONSTANTS.bounds,
         equipment: {
-          readyItemCost: CONSTANTS.rules.equipment.readyItemCost,
+          readyItemCost: {
+            direct:
+              this.actor.data.data.equipment.equipActionCosts.readyDirect.total,
+            fromSlot:
+              this.actor.data.data.equipment.equipActionCosts.readyFromSlot
+                .total
+          },
           readiedItem,
           weaponSlots: this.actor.weaponSlotWeapons.map((weapon) =>
             weapon ? this.toSheetWeapon(weapon) : null
@@ -267,6 +281,8 @@ export default class WvActorSheet extends ActorSheet {
           equipment: HANDLEBARS.partPaths.actor.equipment,
           header: HANDLEBARS.partPaths.actor.header,
           inventory: HANDLEBARS.partPaths.actor.inventory,
+          inventoryEquipmentSlot:
+            HANDLEBARS.partPaths.actor.inventoryEquipmentSlot,
           magic: HANDLEBARS.partPaths.actor.magic,
           stats: HANDLEBARS.partPaths.actor.stats,
           weaponSlot: HANDLEBARS.partPaths.actor.weaponSlot
@@ -641,6 +657,35 @@ export default class WvActorSheet extends ActorSheet {
     attack.execute();
   }
 
+  /** Handle a click event on an unequip item button. */
+  protected async onClickUnequipItem(event: Event): Promise<void> {
+    if (!(event.target instanceof HTMLElement))
+      throw new Error("The target was not an HTMLElement.");
+
+    const itemElement = event.target.closest("[data-equipment-slot]");
+    if (!(itemElement instanceof HTMLElement))
+      throw new Error("The item element parent is not an HTMLElement.");
+
+    const slotName = itemElement.dataset.equipmentSlot;
+    if (typeof slotName !== "string" || !isEquipmentSlot(slotName)) return;
+
+    try {
+      if (slotName === "readiedItem") {
+        await this.actor.unreadyItem();
+      } else if (slotName === "weaponSlot1") {
+        await this.actor.unslotWeapon(1);
+      } else if (slotName === "weaponSlot2") {
+        await this.actor.unslotWeapon(2);
+      } else if (isApparelSlot(slotName)) {
+        await this.actor.unequipApparel(slotName);
+      }
+    } catch (e) {
+      if (e instanceof SystemRulesError && e.key) {
+        ui.notifications?.error(e.key, { localize: true });
+      }
+    }
+  }
+
   /** Handle a click event on a create item button. */
   protected async onClickCreateItem(event: Event): Promise<void> {
     if (!(event.target instanceof HTMLElement))
@@ -829,7 +874,10 @@ export default class WvActorSheet extends ActorSheet {
       if (dragData.data._id !== this.actor.readiedItem?.id) {
         slotsToAllow.push("readiedItem");
 
-        if (this.actor.inCombat) {
+        if (
+          this.actor.inCombat &&
+          this.actor.data.data.equipment.quickSlots.value > 0
+        ) {
           this.markReadySlot("quick-slotable");
         }
       }
@@ -1014,6 +1062,7 @@ interface SheetData extends ActorSheet.Data {
       equipment: string;
       header: string;
       inventory: string;
+      inventoryEquipmentSlot: string;
       magic: string;
       stats: string;
       weaponSlot: string;