+++ src/main/typescript/data/actor/character/equipment/properties.ts
@@ -7,12 +7,16 @@ export default class EquipmentProperties extends EquipmentSource {
     super();
     foundry.utils.mergeObject(this, source);
     this.quickSlots = CompositeResource.from(source.quickSlots);
+    this.quickSlots.source = 0;
   }
 
   override quickSlots: CompositeResource;
 
   /** The damage threshold of the character */
   damageThreshold = new CompositeNumber();
+
+  /** AP costs for various equip actions */
+  equipActionCosts = new EquipActionCosts();
 
   /**
    * Modify the damage threshold and max quick slots by the equipped apparel's
@@ -26,11 +30,22 @@ export default class EquipmentProperties extends EquipmentSource {
           labelComponents: [{ text: apparel.name ?? "" }]
         });
 
-      if (apparel.data.data.quickSlots)
+      if (apparel.data.data.quickSlots.total)
         this.quickSlots.add({
           value: apparel.data.data.quickSlots.total,
           labelComponents: [{ text: apparel.name ?? "" }]
         });
     });
   }
+}
+
+export class EquipActionCosts {
+  /** The cost for unreadying an item or weapon */
+  unready = new CompositeNumber(0);
+
+  /** The cost for readying an item from the inventory */
+  readyDirect = new CompositeNumber(8);
+
+  /** The cost for readying an item with a slot */
+  readyFromSlot = new CompositeNumber(3);
 }