+++ CHANGELOG.md
@@ -14,6 +14,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - an explanation to game master movement settings, that they only affect the
   current user and only if they are a game master
   ([#452](https://github.com/Wasteland-Ventures-Group/WV-VTT-module/issues/452))
+- ability to edit equipped items from equipment slots in inventory
+  ([#483](https://github.com/Wasteland-Ventures-Group/WV-VTT-module/issues/483))
+- ability to unequip apparel from equipment slots
+  ([#483](https://github.com/Wasteland-Ventures-Group/WV-VTT-module/issues/483))
+- ability to unready weapons and items in and out of combat
+  ([#483](https://github.com/Wasteland-Ventures-Group/WV-VTT-module/issues/483))
+- ability to unslot weapons from weapon slots out of combat
+  ([#483](https://github.com/Wasteland-Ventures-Group/WV-VTT-module/issues/483))
+- ability edit current quick slot charges on equipment tab
+  ([#483](https://github.com/Wasteland-Ventures-Group/WV-VTT-module/issues/483))
+- quick slot display in the inventory
+  ([#483](https://github.com/Wasteland-Ventures-Group/WV-VTT-module/issues/483))
+- automated subtraction of quick slot charges when readying with quick slot
+  ([#483](https://github.com/Wasteland-Ventures-Group/WV-VTT-module/issues/483))
+- automated quick slots restore when actor leaves last combat they were in
+  ([#483](https://github.com/Wasteland-Ventures-Group/WV-VTT-module/issues/483))
+- automated action points restore when actor leaves last combat they were in
 
 ### Changed
 
@@ -25,11 +42,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#384](https://github.com/Wasteland-Ventures-Group/WV-VTT-module/issues/384))
 - hard-coded pain threshold table to dynamic display of current pain threshold
   ([#455](https://github.com/Wasteland-Ventures-Group/WV-VTT-module/issues/455))
+- trying to equip an apparel with blocked slots will now fail if one of the
+  blocked slots is already occupied
+  ([#483](https://github.com/Wasteland-Ventures-Group/WV-VTT-module/issues/483))
+- ready item costs to properties tracked on characters, instead of hard-coded
+  values
+  ([#483](https://github.com/Wasteland-Ventures-Group/WV-VTT-module/issues/483))
 
 ### Fixed
 
 - a typo in the GM movement settings
   ([#451](https://github.com/Wasteland-Ventures-Group/WV-VTT-module/issues/451))
+- the apparel sheet unchecking blocked apparel slots
 
 ## [0.20.0] - 2022-08-28
 
@@ -106,7 +130,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - show size category on the statistics tab on the character sheet
 - make tags editable on attacks and ranges
 - show tags on equipment tab
-- equipped items can now be edited directly from the equipment page
+- equipped items can now be edited directly from the equipment tab
 - modifiable initiative
   ([#383](https://github.com/Wasteland-Ventures-Group/WV-VTT-module/issues/383))
 - make SPECIAL and skill rolls always ask for a modifier and add a checkbox to
@@ -190,8 +214,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   modifiers
 - item sheets now show their final stats on the stats tab and allow editing on
   the source values tab
-- moved damage threshold and quick slots max to statistics page
-- expanded the criticals block on statistics page to hold other statistics
+- moved damage threshold and quick slots max to statistics tab
+- expanded the criticals block on statistics tab to hold other statistics
 
 ### Fixed
 
@@ -211,7 +235,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- grid units to the range display on the equipment page
+- grid units to the range display on the equipment tab
 
 ### Changed
 
@@ -285,7 +309,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - actors are now only considered in combat, if they are part of a started combat
   in the active scene or part of an unlinked, started combat
   ([#210](https://github.com/Wasteland-Ventures-Group/WV-VTT-module/issues/210))
-- the way the ready item cost is displayed on the equipment and inventory pages
+- the way the ready item cost is displayed on the equipment and inventory tabs
   ([#216](https://github.com/Wasteland-Ventures-Group/WV-VTT-module/issues/216))
 - the system now tracks per item, whether their data should stay linked to their
   compendium entry
@@ -311,7 +335,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#193](https://github.com/Wasteland-Ventures-Group/WV-VTT-module/issues/193))
 - basic support for apparel items
   ([#140](https://github.com/Wasteland-Ventures-Group/WV-VTT-module/issues/140))
-- rework equpment page to show weapon slots and readied item
+- rework equpment tab to show weapon slots and readied item
   ([#192](https://github.com/Wasteland-Ventures-Group/WV-VTT-module/issues/192))
 
 ### Changed